services/predictor/pyproject.toml
[project]
name = "predictor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Karim Abousselham", email = "karim.abousselham@gmail.com" }
]
requires-python = ">=3.12.11"
dependencies = [
    "great-expectations>=0.18.8",
    "lazypredict>=0.2.16",
    "mlflow>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "risingwave-py>=0.0.2",
    "scikit-learn>=1.7.1",
    "ydata-profiling>=4.16.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
